+++ CHANGELOG.md
@@ -10,6 +10,7 @@ All notable changes to Rainbow-Node-SDK will be documented in this file.
 -   fix  RESTService::retrieveFileDescriptors method, the type of the fileName.
 -   Update for RQRAINB-7234 : NodeJS [Edge] Meet Bubble link should work also on Edges.
 -   Get user's resources presences information from server.
+-   Fix type compare of array when finding DNS.
 
 ## [2.15.1] - 2022-09-02
 -   Updated discover of HTTPoverXMPP to use a message stanza instead of presence one.